+++ src/x2net/Links/Socket/AbstractTcpClient.cs
@@ -173,14 +173,6 @@ namespace x2net
         }
 
         /// <summary>
-        /// Synchronously connects to the endpoint RemoteHost:RemotePort.
-        /// </summary>
-        public void ConnectSync()
-        {
-            ConnectSync(RemoteHost, RemotePort);
-        }
-
-        /// <summary>
         /// Connects to the specified remote address (host:port).
         /// </summary>
         public void Connect(string address)
@@ -189,35 +181,10 @@ namespace x2net
         }
 
         /// <summary>
-        /// Synchronously connects to the specified remote address (host:port).
-        /// </summary>
-        public void ConnectSync(string address)
-        {
-            ConnectSync(address, 0);
-        }
-
-        /// <summary>
         /// Connects to the specified remote address. The port parameter is used
         /// when the given address does not contain a remote port number,
         /// </summary>
         public void Connect(string address, int port)
-        {
-            TryParseAddress(address, port);
-            Connect(HostToIPAddress(RemoteHost), RemotePort);
-        }
-
-        /// <summary>
-        /// Synchronously connects to the specified remote address. The por
-        /// parameter is used when the given address does not contain a remote
-        /// port number,
-        /// </summary>
-        public void ConnectSync(string address, int port)
-        {
-            TryParseAddress(address, port);
-            ConnectSync(HostToIPAddress(RemoteHost), RemotePort);
-        }
-
-        private void TryParseAddress(string address, int port)
         {
             int index = address.LastIndexOf(':');
             if (index >= 0 && address.Length > 15)
@@ -240,20 +207,20 @@ namespace x2net
 
             RemoteHost = address;
             RemotePort = port;
-        }
 
-        private IPAddress HostToIPAddress(string host)
-        {
+            IPAddress ip;
             try
             {
-                return Dns.GetHostAddresses(host)[0];
+                ip = Dns.GetHostAddresses(address)[0];
             }
             catch (Exception e)
             {
                 Trace.Error("{0} error resolving target host {1} : {2}",
-                    Name, host, e.Message);
+                    Name, address, e.Message);
                 throw;
             }
+
+            Connect(ip, RemotePort);
         }
 
         public void ConnectAndSend(Event e)
@@ -347,23 +314,6 @@ namespace x2net
             Connect(null, new IPEndPoint(ip, port));
         }
 
-
-        /// <summary>
-        /// Synchronously connects to the specified IP address and port.
-        /// </summary>
-        public void ConnectSync(IPAddress ip, int port)
-        {
-            connecting = true;
-
-            if (session != null &&
-                ((AbstractTcpSession)session).SocketConnected)
-            {
-                throw new InvalidOperationException();
-            }
-
-            ConnectSync(null, new IPEndPoint(ip, port));
-        }
-
         public void Disconnect()
         {
             LinkSession currentSession = Session;
@@ -391,15 +341,6 @@ namespace x2net
             startTime = DateTime.UtcNow;
 
             ConnectInternal(socket, endpoint);
-        }
-
-        private void ConnectSync(Socket socket, EndPoint endpoint)
-        {
-            Trace.Info("{0} synchronously connecting to {1}", Name, endpoint);
-
-            startTime = DateTime.UtcNow;
-
-            ConnectInternalSync(socket, endpoint);
         }
 
         // Reconnects to the last successful remote address.
@@ -492,12 +433,6 @@ namespace x2net
         /// Provides an actual implementation of asynchronous Connect.
         /// </summary>
         protected abstract void ConnectInternal(Socket socket, EndPoint endpoint);
-
-
-        /// <summary>
-        /// Provides an actual implementation of synchronous Connect.
-        /// </summary>
-        protected abstract bool ConnectInternalSync(Socket socket, EndPoint endpoint);
 
         /// <summary>
         /// <see cref="ClientLink.OnConnectInternal"/>